+++ packages/terra-layout/CHANGELOG.md
@@ -10,6 +10,9 @@ Unreleased
 * Debounce to resize of layout
 * Use display none on the hover section when hidden.
 
+###Changed
+* Updating tests to point to new terra-framework-site location.
+
 1.0.0 - (November 17, 2017)
 ------------------
 Initial stable release